Fixed Matrix Issue!
Implemented a simple set of ImGui controls;
After a lot of debugging, I found that the whole issue with my matrices was simple: DirectX 11 matrices a Row Major format, and a Left-Handed coordinate system. However, I was passing in a Column Major matrix, A Simple Transposition fixed all of my issues. 

Now, I just have to tidy things up a little, and fix the draw order of my geometry.